Reduce TCP send buffer from 64k to 0
This _might_ reduce latency in situations where the receiver is unable to keep up with the sender, and thus resulting in server-side buildup of encoded frames in the TCP send buffer. Sending of new frames to the socket will now instead be blocked until they can be transfered over the wire.

WARNING: I've not actually verified that it helps.

DOC: https://docs.microsoft.com/en-us/troubleshoot/windows/win32/data-segment-tcp-winsock